% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{pmx_priors}
\alias{pmx_priors}
\title{Collect public priors for the released corrections}
\usage{
pmx_priors(...)
}
\arguments{
\item{...}{Named \code{\link[=pmx_prior]{pmx_prior()}} objects. Recognized names are \code{pk} and \code{pd}.}
}
\value{
A \code{pmx_priors} object.
}
\description{
Collect public priors for the released corrections
}
